Refactored settings provider for more security against data loss
Settings Provider (SP) was using LocalStorage (with Ionic's abstraction) and it could eventually lead to data losses if storage was to be missing on user's device (iOS). With this commit, the SP begins using an hybrid storage solution : LocalStorage for fast storage and, if this is empty, it checks in PouchDB to see if there's a backup so there's no data loss. If the backup is empty, it fills in the defaults. Language Provider now uses an async method to get the setting at startup to account for PouchDB's slowness.